requirements.txt
python-telegram-bot[job-queue] --pre
audiofile
librosa
pdfrw
requests
numpy
scipy
Pillow
transformers
torch
openai
together